Revert "feat(plugin) Make E2E tests independent over unit tests (#695)"
This reverts commit 29267ce364f641e897544168fee979546cf8c454.
