+++ e2e/plugin/fixtures/fixtures.go
@@ -4,58 +4,100 @@
 package fixtures
 
 import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 
-	"github.com/cloudoperators/greenhouse/e2e/shared"
 	greenhousev1alpha1 "github.com/cloudoperators/greenhouse/pkg/apis/greenhouse/v1alpha1"
+	"github.com/cloudoperators/greenhouse/pkg/test"
 )
 
+func PreparePluginDefinition(name, namespace string, opts ...func(*greenhousev1alpha1.PluginDefinition)) *greenhousev1alpha1.PluginDefinition {
+	pd := &greenhousev1alpha1.PluginDefinition{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "PluginDefinition",
+			APIVersion: greenhousev1alpha1.GroupVersion.String(),
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: namespace,
+		},
+		Spec: greenhousev1alpha1.PluginDefinitionSpec{
+			Description: "TestPluginDefinition",
+			Version:     "1.0.0",
+			HelmChart:   &greenhousev1alpha1.HelmChartReference{}, // helm chart values are override later
+		},
+	}
+	for _, o := range opts {
+		o(pd)
+	}
+
+	return pd
+}
+
 func PrepareNginxPluginDefinition(namespace string) *greenhousev1alpha1.PluginDefinition {
-	return shared.PreparePluginDefinition("nginx-18.1.7", namespace,
-		shared.WithVersion("18.1.7"),
-		shared.WithDescription("TestPluginDefinition"),
-		shared.WithHelmChart(
+	return PreparePluginDefinition("nginx-18.1.7", namespace,
+		test.WithVersion("18.1.7"),
+		test.WithHelmChart(
 			&greenhousev1alpha1.HelmChartReference{
 				Name:       "bitnamicharts/nginx",
 				Repository: "oci://registry-1.docker.io",
 				Version:    "18.1.7",
 			},
 		),
-		shared.AppendPluginOption(greenhousev1alpha1.PluginOption{
+		test.AppendPluginOption(greenhousev1alpha1.PluginOption{
 			Default:     &apiextensionsv1.JSON{Raw: []byte("false")},
 			Description: "autoscaling.enabled",
 			Name:        "autoscaling.enabled",
 			Type:        "bool",
 		}),
-		shared.AppendPluginOption(greenhousev1alpha1.PluginOption{
+		test.AppendPluginOption(greenhousev1alpha1.PluginOption{
 			Default:     &apiextensionsv1.JSON{Raw: []byte("\"\"")},
 			Description: "autoscaling.maxReplicas",
 			Name:        "autoscaling.maxReplicas",
 			Type:        "string",
 		}),
-		shared.AppendPluginOption(greenhousev1alpha1.PluginOption{
+		test.AppendPluginOption(greenhousev1alpha1.PluginOption{
 			Default:     &apiextensionsv1.JSON{Raw: []byte("\"\"")},
 			Description: "autoscaling.minReplicas",
 			Name:        "autoscaling.minReplicas",
 			Type:        "string",
 		}),
-		shared.AppendPluginOption(greenhousev1alpha1.PluginOption{
+		test.AppendPluginOption(greenhousev1alpha1.PluginOption{
 			Default:     &apiextensionsv1.JSON{Raw: []byte("true")},
 			Description: "containerSecurityContext.enabled",
 			Name:        "containerSecurityContext.enabled",
 			Type:        "bool",
 		}),
-		shared.AppendPluginOption(greenhousev1alpha1.PluginOption{
+		test.AppendPluginOption(greenhousev1alpha1.PluginOption{
 			Default:     &apiextensionsv1.JSON{Raw: []byte("1")},
 			Description: "replicaCount",
 			Name:        "replicaCount",
 			Type:        "int",
 		}),
-		shared.AppendPluginOption(greenhousev1alpha1.PluginOption{
+		test.AppendPluginOption(greenhousev1alpha1.PluginOption{
 			Default:     &apiextensionsv1.JSON{Raw: []byte("true")},
 			Description: "podSecurityContext.enabled",
 			Name:        "podSecurityContext.enabled",
 			Type:        "bool",
 		}),
 	)
+}
+
+func PreparePlugin(name, namespace string, opts ...func(*greenhousev1alpha1.Plugin)) *greenhousev1alpha1.Plugin {
+	plugin := &greenhousev1alpha1.Plugin{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "Plugin",
+			APIVersion: greenhousev1alpha1.GroupVersion.String(),
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:         name,
+			Namespace:    namespace,
+			GenerateName: name + "-gen",
+		},
+	}
+	for _, o := range opts {
+		o(plugin)
+	}
+	return plugin
 }
